+++ eth_vertigo/mutator/source_mutator.py
@@ -1,4 +1,4 @@
-from eth_vertigo.source.source_file import SourceFile
+from eth_vertigo.mutator.source_file import SourceFile
 from pathlib import Path
